+++ docs/conf.py
@@ -25,9 +25,9 @@ project = 'moban'
 copyright = '2017-2019 Onni Software Ltd.'
 author = 'C. W.'
 # The short X.Y version
-version = '0.6.0'
+version = '0.6.1'
 # The full version, including alpha/beta/rc tags
-release = '0.5.0'
+release = '0.6.1'
 
 # -- General configuration ---------------------------------------------------