+++ moban/constants.py
@@ -36,6 +36,8 @@ LABEL_MOBANFILE = "mobanfile"
 LABEL_FORCE = "force"
 LABEL_VERSION = "version"
 LABEL_GROUP = "group"
+LABEL_DEFINE = "define"
+CLI_DICT = "user_dict"
 
 DEFAULT_CONFIGURATION_DIRNAME = ".%s.cd" % PROGRAM_NAME
 DEFAULT_TEMPLATE_DIRNAME = ".%s.td" % PROGRAM_NAME
@@ -70,7 +72,7 @@ LABEL_SOURCE = "source"
 LABEL_DEST = "destination"
 LABEL_FORCE_TEMPLATE_TYPE = "force_template_type"
 LABEL_TEMPLATE_TYPES = "template_types"
-LABEL_DEBUG = "debug"
+LABEL_VERBOSE = "verbose"
 
 # error messages
 ERROR_DATA_FILE_NOT_FOUND = "Both %s and %s does not exist"