+++ moban/main.py
@@ -24,9 +24,11 @@ from moban import (
 )
 from moban._version import __version__
 from moban.hashstore import HASH_STORE
+from moban.program_options import OPTIONS
 from moban.data_loaders.manager import merge, load_data
 
 LOG = logging.getLogger()
+LOG_LEVEL = [logging.WARNING, logging.INFO, logging.DEBUG]
 
 
 def main():
@@ -36,11 +38,11 @@ def main():
     parser = create_parser()
     options = vars(parser.parse_args())
     HASH_STORE.IGNORE_CACHE_FILE = options[constants.LABEL_FORCE]
-    if options[constants.LABEL_DEBUG]:
-        logging.basicConfig(
-            format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
-            level=logging.DEBUG,
-        )
+    options[constants.CLI_DICT] = handle_custom_variables(
+        options.pop(constants.LABEL_DEFINE)
+    )
+    OPTIONS.update(options)
+    handle_verbose(options[constants.LABEL_VERBOSE])
 
     moban_file = options[constants.LABEL_MOBANFILE]
     load_engine_factory_and_engines()  # Error: jinja2 if removed
@@ -135,17 +137,23 @@ def create_parser():
         help="string templates",
     )
     parser.add_argument(
-        "-v",
+        "-V",
         "--%s" % constants.LABEL_VERSION,
         action="version",
         version="%(prog)s {v}".format(v=__version__),
     )
     parser.add_argument(
+        "-v",
+        action="count",
+        dest=constants.LABEL_VERBOSE,
+        default=0,
+        help="show verbose",
+    )
+    parser.add_argument(
         "-d",
-        action="store_true",
-        dest=constants.LABEL_DEBUG,
-        default=False,
-        help="to show debug trace",
+        "--%s" % constants.LABEL_DEFINE,
+        nargs="+",
+        help="to take a list of VAR=VALUEs",
     )
     return parser
 
@@ -245,3 +253,23 @@ def find_default_moban_file():
 
 def load_engine_factory_and_engines():
     plugins.make_sure_all_pkg_are_loaded()
+
+
+def handle_custom_variables(list_of_definitions):
+    custom_data = {}
+    if list_of_definitions:
+        for definition in list_of_definitions:
+            key, value = definition.split("=")
+            custom_data[key] = value
+
+    return custom_data
+
+
+def handle_verbose(verbose_level):
+    if verbose_level > len(LOG_LEVEL):
+        verbose_level = 3
+    level = LOG_LEVEL[verbose_level]
+    logging.basicConfig(
+        format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+        level=level,
+    )