+++ moban/program_options.py
@@ -1,0 +1,3 @@
+from moban import constants
+
+OPTIONS = {constants.CLI_DICT: {}}